refactor: extract matchLabelText
