+++ src/helpers/matchers/matchTextContent.ts
@@ -1,10 +1,6 @@
-import { Text } from 'react-native';
 import type { ReactTestInstance } from 'react-test-renderer';
-import { getConfig } from '../../config';
 import { matches, TextMatch, TextMatchOptions } from '../../matches';
-import { filterNodeByType } from '../filterNodeByType';
 import { getTextContent } from '../getTextContent';
-import { getHostComponentNames } from '../host-component-names';
 
 /**
  * Matches the given node's text content against string or regex matcher.
@@ -18,13 +14,6 @@ export function matchTextContent(
   text: TextMatch,
   options: TextMatchOptions = {}
 ) {
-  const textType = getConfig().useBreakingChanges
-    ? getHostComponentNames().text
-    : Text;
-  if (!filterNodeByType(node, textType)) {
-    return false;
-  }
-
   const textContent = getTextContent(node);
   const { exact, normalizer } = options;
   return matches(text, textContent, normalizer, exact);
