+++ src/queries/labelText.ts
@@ -1,7 +1,7 @@
 import type { ReactTestInstance } from 'react-test-renderer';
 import { findAll } from '../helpers/findAll';
-import { matches, TextMatch, TextMatchOptions } from '../matches';
-import { matchTextContent } from '../helpers/matchers/matchTextContent';
+import { TextMatch, TextMatchOptions } from '../matches';
+import { matchLabelText } from '../helpers/matchers/matchLabelText';
 import { makeQueries } from './makeQueries';
 import type {
   FindAllByQuery,
@@ -15,66 +15,17 @@ import { CommonQueryOptions } from './options';
 
 type ByLabelTextOptions = CommonQueryOptions & TextMatchOptions;
 
-const getNodeByLabelText = (
-  node: ReactTestInstance,
-  text: TextMatch,
-  options: TextMatchOptions
-) => {
-  const { exact, normalizer } = options;
-  return matches(text, node.props.accessibilityLabel, normalizer, exact);
-};
-
-const matchAccessibilityLabelledByText = (
-  rootInstance: ReactTestInstance,
-  nativeID: string | undefined,
-  text: TextMatch,
-  options: TextMatchOptions
-) => {
-  if (!nativeID) {
-    return false;
-  }
-
-  const { exact, normalizer } = options;
-
-  return rootInstance
-    .findAll((node) =>
-      matches(nativeID, node.props?.nativeID, normalizer, exact)
-    )
-    .some((node) => node.findAll((n) => matchTextContent(n, text)).length > 0);
-};
-
-const matchAccessibilityLabel = (
-  rootInstance: ReactTestInstance,
-  node: ReactTestInstance,
-  text: TextMatch,
-  options: TextMatchOptions = {}
-) => {
-  return (
-    getNodeByLabelText(node, text, options) ||
-    matchAccessibilityLabelledByText(
-      rootInstance,
-      node.props.accessibilityLabelledBy,
-      text,
-      options
-    )
-  );
-};
-
-const queryAllByLabelText = (
-  instance: ReactTestInstance
-): ((
-  text: TextMatch,
-  queryOptions?: ByLabelTextOptions
-) => Array<ReactTestInstance>) =>
-  function queryAllByLabelTextFn(text, queryOptions) {
+function queryAllByLabelText(instance: ReactTestInstance) {
+  return (text: TextMatch, queryOptions?: ByLabelTextOptions) => {
     return findAll(
       instance,
       (node) =>
         typeof node.type === 'string' &&
-        matchAccessibilityLabel(instance, node, text, queryOptions),
+        matchLabelText(instance, node, text, queryOptions),
       queryOptions
     );
   };
+}
 
 const getMultipleError = (labelText: TextMatch) =>
   `Found multiple elements with accessibilityLabel: ${String(labelText)} `;
